perf(#3225): split the §2 analyzer's rendering out (campsite rule, ~800-line target)
analyze-3225.py had grown to 1094 lines. Rendering is pure presentation over the
analysis dict and derives no numbers, so it moves to analyze_3225_render.py
(842 + 264). Verified behaviour-preserving: the emitted analysis is identical to
the committed results/analysis-3225.txt apart from the results-dir line, and
--smoke still reproduces 29/29 of #3217's published medians. Dropped the now-unused
`re` import.
